requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
discord.py==2.3.2
frozenlist==1.4.0
idna==3.4
joblib==1.3.2
multidict==6.0.4
nltk==3.8.1
regex==2023.10.3
tqdm==4.66.1
yarl==1.9.2
